Split runAssignment test to multiple files
