+++ server/src/features/feedback/feedbackController.test.ts
@@ -16,6 +16,7 @@ beforeEach(async () => {
   server = await startServer({
     dbConnString: mongoServer.getUri(),
     dbName: faker.random.alphaNumeric(10),
+    enableSentry: false,
   });
 });
