+++ server/src/features/player-assignment/runAssignment.test.ts
@@ -66,216 +66,6 @@ describe("Assignment with valid data", () => {
     );
   });
 
-  test("should return success with group strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    // FIRST RUN
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("success");
-    expect(assignResults.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults = assignResults.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-
-    // SECOND RUN
-
-    const assignResults2 = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults2.status).toEqual("success");
-    expect(assignResults2.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults2 = assignResults2.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults2.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-  });
-
-  test("should return success with padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    // FIRST RUN
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("success");
-    expect(assignResults.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults = assignResults.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-
-    // SECOND RUN
-
-    const assignResults2 = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults2.status).toEqual("success");
-    expect(assignResults2.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults2 = assignResults2.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults2.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-  });
-
-  test("should return success with random strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    // FIRST RUN
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("success");
-    expect(assignResults.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults = assignResults.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-
-    // SECOND RUN
-
-    const assignResults2 = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults2.status).toEqual("success");
-    expect(assignResults2.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults2 = assignResults2.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults2.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-  });
-
-  test("should return success with group+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    // FIRST RUN
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("success");
-    expect(assignResults.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults = assignResults.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-
-    // SECOND RUN
-
-    const assignResults2 = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults2.status).toEqual("success");
-    expect(assignResults2.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults2 = assignResults2.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults2.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-  });
-
-  test("should return success with random+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    // FIRST RUN
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("success");
-    expect(assignResults.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults = assignResults.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-
-    // SECOND RUN
-
-    const assignResults2 = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults2.status).toEqual("success");
-    expect(assignResults2.results.length).toBeGreaterThanOrEqual(
-      expectedResultsCount
-    );
-
-    const groupResults2 = assignResults2.results.filter((result) =>
-      groupTestUsers.includes(result.username)
-    );
-    expect(groupResults2.length).toBeOneOf([groupTestUsers.length, 0]);
-
-    await verifyUserSignups();
-  });
-
   test("should return valid results after multiple executions on different times", async () => {
     const { CONVENTION_START_TIME } = sharedConfig;
     const assignmentStrategy = AssignmentStrategy.PADG;
@@ -642,161 +432,5 @@ describe("Assignment with first time bonus", () => {
     expect(
       previousDirectSignupAlwaysOpenSignup?.userSignups[0].username
     ).toEqual(mockUser2.username);
-  });
-});
-
-describe("Assignment with no games", () => {
-  beforeEach(async () => {
-    const newUsersCount = 1;
-    const groupSize = 0;
-    const numberOfGroups = 0;
-    const newGamesCount = 0;
-    const testUsersCount = 0;
-
-    await generateTestData(
-      newUsersCount,
-      newGamesCount,
-      groupSize,
-      numberOfGroups,
-      testUsersCount
-    );
-  });
-
-  test("should return error with group strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no starting games");
-  });
-
-  test("should return error with padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no starting games");
-  });
-
-  test("should return error with random strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no starting games");
-  });
-
-  test("should return error with group+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no starting games");
-  });
-
-  test("should return error with random+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no starting games");
-  });
-});
-
-describe("Assignment with no players", () => {
-  beforeEach(async () => {
-    const newUsersCount = 0;
-    const groupSize = 0;
-    const numberOfGroups = 0;
-    const newGamesCount = 1;
-    const testUsersCount = 0;
-
-    await generateTestData(
-      newUsersCount,
-      newGamesCount,
-      groupSize,
-      numberOfGroups,
-      testUsersCount
-    );
-  });
-
-  test("should return error with group strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no signup wishes");
-  });
-
-  test("should return error with padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no signup wishes");
-  });
-
-  test("should return error with random strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no signup wishes");
-  });
-
-  test("should return error with group+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.GROUP_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no signup wishes");
-  });
-
-  test("should return error with random+padg strategy", async () => {
-    const { CONVENTION_START_TIME } = sharedConfig;
-    const assignmentStrategy = AssignmentStrategy.RANDOM_PADG;
-    const startingTime = dayjs(CONVENTION_START_TIME).add(2, "hours").format();
-
-    const assignResults = await runAssignment({
-      assignmentStrategy,
-      startingTime,
-    });
-    expect(assignResults.status).toEqual("error: no signup wishes");
   });
 });
