+++ server/src/features/user/signup-message/signupMessageController.test.ts
@@ -29,6 +29,7 @@ beforeEach(async () => {
   server = await startServer({
     dbConnString: mongoServer.getUri(),
     dbName: faker.random.alphaNumeric(10),
+    enableSentry: false,
   });
 });
