+++ server/src/utils/server.test.ts
@@ -15,6 +15,7 @@ beforeEach(async () => {
   server = await startServer({
     dbConnString: mongoServer.getUri(),
     dbName: faker.random.alphaNumeric(10),
+    enableSentry: false,
   });
 });
